fix(core): unify frontend and execution contracts
Resolve binding, initialization, scheduling, GC and native recovery defects
across the compiler and execution backends.

- Preserve checked call arguments and declaration identities, split checker
  responsibilities, and centralize package analysis and physical layouts.
- Account for fiber auxiliary storage, bound scheduler and GC work, and align
  defer, unwind, select and asynchronous execution across VM and Wasm.
- Correct JIT heap effects and scalar recovery lifetimes, constrain inlining,
  and share native stack guards, scratch storage and artifact validation.
- Add bounded polymorphic call caches, outline repeated helper paths, and
  reduce constant expansion, cold analysis and unused recovery metadata.
- Add frontend, runtime and optimizer regressions, prove optimized execution,
  and update backend contracts and generated documentation.

Native AOT ABI advances to version 2; existing native artifacts must be rebuilt.

Validation includes 3,559 language cases, native crate tests, focused optimized
JIT/GC tests, and actual no_std, Wasm VM, Wasm AOT and Native AOT execution.
Formatting and full repository lint pass on the staged source snapshot.
The 21-case benchmark comparison retains a confirmed Fibonacci regression in
JIT and Native AOT; historical timing changes also include host differences.

+++ lang/crates/vo-wasm-aot/src/codegen/frame/interface.rs
@@ -214,7 +214,9 @@ pub(super) fn compile(
                     .instruction(&W::LocalSet(ALLOC_LOCAL));
                 if array_layout.elem_bytes < 8 {
                     for index in 0..array_layout.len {
-                        store_prefix(body, instruction.a + index);
+                        // The assertion helper checked the complete result
+                        // width against the u16 frame ABI before this loop.
+                        store_prefix(body, instruction.a + index as u16);
                         body.instruction(&W::LocalGet(ALLOC_LOCAL));
                         let offset = u64::from(index) * u64::from(array_layout.elem_bytes);
                         match (array_layout.elem_bytes, array_layout.needs_sign_extend) {
